Adds the recurrences for the events
- adds django-recurrence app
- adds recurrence model and form
- displays recurrence form in the create event view
- create the recurrence object

Since we need the recurrence only on recurring events, we need to
display the form only if the recurrence option of the event form is
activated (solved in js).
We don't need the option to set an end to the recurrence (the limits are
handled by the event time model and its form). This option is hidden
using css.
To keep things simple, we don't want the first rule to be an exclusion
rule. For this reason, the exclusion rule option is hidden from the
first block of rules. This also requires the first rule to be permanent
(meaning, you can't remove it) otherwise the following situation can
happen:
1. the user adds the first rule (the exclusion button is not displayed)
2. the user adds a second rule (and checks the exclusion)
3. the user removes the first rule
the second rules then becomes the first rule and therefore the exclusion
option is hidden (in css).