Allows to run actions on event lists
When listing events and proposals, it is handy to be able to
publish/hide/delete several of them at the same time. This adds an
editorial view which allows to select several events with the use of a
checkbox and allows to perform actions on them depending on their
publication status.