Moves gis solution to custom file
Since the gis solution is very dependent of the project (not everyone
wants to use the google static api), it is necessary to have a plug-able
solution. By moving the inclusion of gis to its own template (which is
included in the event details view), we allow easy customization by
template overloading.

Also add the additional address information