fixes filter form mixin for list in get parameters
a list sent as GET parameters is commonly sent like this:
    `?key[]=value1&key[]=value2&key[]=value3`

the braces in the keyword name break the logic of the django field.
a simple way to work around this is to prepare the data before creating
the form, removing all the braces from the keyword names.
this fix makes sure that the filters are passed as list with adapted
names to match the expectations of the django field.