Fix prefilled values after page navigation
When using the page navigation, get parameters are set.
In that case the FilterForm got a non-empty request.GET, such that
prefilled fields' initial values did not appear.
The initial values need to be added to the data passed to the FilterForm
if a non-empty request.GET is received.

...and while we're at it, some of the FilterFormMixin's code was
refactored.