Improve title on event page for Solr indexing
If the title of the calendar is displayed first, the title of the event
does not show up in the Solr list of results (because the combination of
both, event and calendar title is too long to be displayed).
By putting the title of the event upfront, the index should still find
the events by the title of the calendar, yet it will allow to display
the title of the event in the search results.

See also:
https://github.com/4teamwork/veranstaltungen_winterthur/issues/132