Fix removal of images
When an images is removed, the value of the field is a boolean (with the
value False).
The error happened when trying to access the image's name attribute,
which does not exist if the image is a boolean.
Therefore an attribute error was raised.
This fixes the error by removing the image when the field is False.