Add option to define the ordering of the filter form fields.
The filter form fields are constructed dynamically based on various Django models, so we need to use orderable data structures when generating them. Otherwise the order of the fields may change when the application is restarted.

Please note that you need to restart the application when you change the order of the groupings in the database.