Fix removal of the event's document.
When the user chooses to remove the event's document, the value of the form field is a boolean (with the value False). This case was not handled at all so it was never possible to remove the document because an exception occurred (attribute error).

The same error has been fixed for the event's image a few years ago (see https://github.com/4teamwork/django-eventary/pull/178).